# apps/task/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findOne",
        "summary": "",
        "description": "Get a Task. The calling user must have Read permission the Task either explicitly or implicitly.\nExplicit Permission: E.g Given by a role e.g developer role\nImplicit Permission: The task criteria matches calling user profile, and calling user have permission to create Feedback.\nImplicit Permission: This user has created Feedback for this Task",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        },
        "tags": ["tasks"]
      },
      "patch": {
        "operationId": "TasksController_update",
        "summary": "",
        "description": "Update a Task. The calling user must have Update permission the Task.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "summary": "",
        "description": "Delete a Task. The calling user must have Delete permission the Task.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/feedbacks": {
      "post": {
        "operationId": "FeedbacksController_create",
        "summary": "",
        "description": "Create a new Feedback. The user must have Create permission for Feedback.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateFeedbackDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Feedback" }
              }
            }
          }
        },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "FeedbacksController_findAll",
        "summary": "",
        "description": "Get a list of all Feedbacks. The user must have permission to Read all Feedbacks.",
        "parameters": [
          {
            "name": "taskId",
            "required": false,
            "in": "query",
            "description": "The task to which the feedbacks belong.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Feedback" }
                }
              }
            }
          }
        },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/feedbacks/search": {
      "get": {
        "operationId": "FeedbacksController_searchAll",
        "summary": "",
        "description": "Get a list of Feedbacks on which the calling user have access to.\nThe user must have Read permission on the Feedbacks.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Feedback" }
                }
              }
            }
          }
        },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/feedbacks/tasks/{taskId}": {
      "get": {
        "operationId": "FeedbacksController_findFeedbacks",
        "summary": "",
        "description": "Get all Feedbacks for a Task. The user must have Read permission on the Task.",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "description": "The task id to list Feedbacks for",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Feedback" }
                }
              }
            }
          }
        },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/feedbacks/{id}": {
      "get": {
        "operationId": "FeedbacksController_findOne",
        "summary": "",
        "description": "Get a Feedback.\nThe calling user must have Read permission on the Feedback.\nThe Read permission for the Feedback is granted if one of the following holds:\n1. The feedback is created by the User.\n2. User have Read permission on the Task to which this Feedback belongs.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Feedback" }
              }
            }
          }
        },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      },
      "patch": {
        "operationId": "FeedbacksController_update",
        "summary": "",
        "description": "Update a Feedback. The calling user must have Update permission on the Feedback.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateFeedbackDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Feedback" }
              }
            }
          }
        },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "FeedbacksController_remove",
        "summary": "",
        "description": "Delete a Feedback. The calling user must have Delete permission on the Feedback.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/feedbacks/{id}/release-payment": {
      "put": {
        "operationId": "FeedbacksController_releasePayment",
        "summary": "",
        "description": "Release payment for the Feedback. User must have Update Permission on the Task related to Feedback.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/feedbacks/{id}/rate-feedback": {
      "patch": {
        "operationId": "FeedbacksController_rateFeedback",
        "summary": "",
        "description": "Rate a feedback. User must have Update Permission on the Task related to Feedback.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RateFeedbackDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Feedback" }
              }
            }
          }
        },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/feedbacks/{id}/rate-task": {
      "patch": {
        "operationId": "FeedbacksController_rateTask",
        "summary": "",
        "description": "Rate a Task related to feedback. User must have Read Permission on the Feedback.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RateTaskDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Feedback" }
              }
            }
          }
        },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": {
    "title": "Task",
    "description": "Task service is responsible for iterative validation of Prototypes. Tasks are created and Feedback is provided through this service.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "Bearer",
        "description": "Authentication is done by a signed JWT",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
...
      "Question": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "description": { "type": "string" },
          "type": {
            "type": "string",
            "enum": ["text", "radio", "star-rating", "thumbs-rating"]
          },
          "order": { "type": "number", "minimum": 1 },
          "radioOptions": { "type": "array", "items": { "type": "string" } },
          "taskId": { "type": "string" }
        },
        "required": [
          "id",
          "description",
          "type",
          "order",
          "radioOptions",
          "taskId"
        ]
      },
...
      "Image": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "prototypeNumber": { "type": "number", "minimum": 1, "maximum": 2 },
          "taskId": { "type": "string" },
          "cloudId": { "type": "string" }
        },
        "required": [
          "id",
          "url",
          "dateCreated",
          "prototypeNumber",
          "taskId",
          "cloudId"
        ]
      },
      "Skill": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" }
        },
        "required": ["id", "name"]
      },
      "Country": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "code": { "type": "string" }
        },
        "required": ["id", "name", "code"]
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "testType": { "enum": ["basic", "comparison"], "type": "string" },
          "prototypeFormat": {
            "enum": ["image", "iframe", "text"],
            "type": "string"
          },
          "status": { "enum": ["draft", "open", "closed"], "type": "string" },
          "accessType": {
            "enum": ["open", "nda", "request"],
            "type": "string"
          },
          "iframeUrl1": { "type": "string" },
          "iframeUrl2": { "type": "string" },
          "textualDescription1": { "type": "string" },
          "textualDescription2": { "type": "string" },
          "minAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "maxAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "minExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "maxExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "budget": { "type": "number" },
          "incentive": { "type": "number" },
          "userId": { "type": "string" },
          "projectId": { "type": "string" },
          "images": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Image" }
          },
          "skills": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Skill" }
          },
          "countries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Country" }
          },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Question" }
          },
          "feedbacks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Feedback" }
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "dateCreated",
          "testType",
          "prototypeFormat",
          "status",
          "accessType",
          "iframeUrl1",
          "iframeUrl2",
          "textualDescription1",
          "textualDescription2",
          "minAge",
          "maxAge",
          "minExperience",
          "maxExperience",
          "budget",
          "incentive",
          "userId",
          "projectId",
          "images",
          "skills",
          "countries",
          "questions",
          "feedbacks"
        ]
      },
      "Answer": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "textAnswer": { "type": "string" },
          "radioAnswer": { "type": "string" },
          "starRatingAnswer": { "type": "number", "minimum": 1, "maximum": 5 },
          "thumbsRatingAnswer": { "type": "string", "enum": ["up", "down"] },
          "feedbackId": { "type": "string" },
          "questionId": { "type": "string" }
        },
        "required": [
          "id",
          "textAnswer",
          "radioAnswer",
          "starRatingAnswer",
          "thumbsRatingAnswer",
          "feedbackId",
          "questionId"
        ]
      },
      "Feedback": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "comment": { "type": "string" },
          "paymentStatus": {
            "enum": ["pending", "completed"],
            "type": "string"
          },
          "feedbackRating": { "type": "number" },
          "feedbackRatingComment": { "type": "string" },
          "taskRating": { "type": "number" },
          "taskRatingComment": { "type": "string" },
          "taskId": { "type": "string" },
          "task": { "$ref": "#/components/schemas/Task" },
          "userId": { "type": "string" },
          "answers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Answer" }
          }
        },
        "required": [
          "id",
          "dateCreated",
          "comment",
          "paymentStatus",
          "feedbackRating",
          "feedbackRatingComment",
          "taskRating",
          "taskRatingComment",
          "taskId",
          "task",
          "userId",
          "answers"
        ]
      },
...
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "testType": { "type": "string", "enum": ["basic", "comparison"] },
          "prototypeFormat": {
            "type": "string",
            "enum": ["image", "iframe", "text"]
          },
          "accessType": {
            "type": "string",
            "enum": ["open", "nda", "request"]
          },
          "iframeUrl1": { "type": "string" },
          "iframeUrl2": { "type": "string" },
          "textualDescription1": { "type": "string" },
          "textualDescription2": { "type": "string" },
          "minAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "maxAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "minExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "maxExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "budget": { "type": "number", "minimum": 0 },
          "incentive": { "type": "number", "minimum": 0 },
          "skills": { "type": "array", "items": { "type": "string" } },
          "countries": { "type": "array", "items": { "type": "string" } },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Question" }
          }
        }
      },
...
      "CreateFeedbackDto": {
        "type": "object",
        "properties": {
          "comment": { "type": "string" },
          "taskId": { "type": "string" },
          "answers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Answer" }
          },
          "taskRating": { "type": "number", "minimum": 1, "maximum": 5 },
          "taskRatingComment": { "type": "string" }
        },
        "required": ["taskId", "answers"]
      },
      "UpdateFeedbackDto": {
        "type": "object",
        "properties": {
          "comment": { "type": "string" },
          "answers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Answer" }
          },
          "taskRating": { "type": "number", "minimum": 1, "maximum": 5 },
          "taskRatingComment": { "type": "string" }
        }
      },
      "RateFeedbackDto": {
        "type": "object",
        "properties": {
          "feedbackRating": { "type": "number", "minimum": 1, "maximum": 5 },
          "feedbackRatingComment": { "type": "string" }
        },
        "required": ["feedbackRating"]
      },
      "RateTaskDto": {
        "type": "object",
        "properties": {
          "taskRating": { "type": "number", "minimum": 1, "maximum": 5 },
          "taskRatingComment": { "type": "string" }
        },
        "required": ["taskRating"]
      }
    }
  }
}